+++ source/Irrlicht/CImageWriterJPG.cpp
@@ -72,7 +72,7 @@ static boolean jpeg_empty_output_buffer(j_compress_ptr cinfo)
 static void jpeg_term_destination(j_compress_ptr cinfo)
 {
 	mem_dest_ptr dest = (mem_dest_ptr) cinfo->dest;
-	const s32 datacount = (s32)(OUTPUT_BUF_SIZE - dest->pub.free_in_buffer);
+	const size_t datacount = OUTPUT_BUF_SIZE - dest->pub.free_in_buffer;
 	// for now just exit upon file error
 	if (dest->file->write(dest->buffer, datacount) != datacount)
 		ERREXIT (cinfo, JERR_FILE_WRITE);
