+++ source/Irrlicht/CMemoryFile.h
@@ -66,7 +66,7 @@ namespace io
 		virtual ~CMemoryWriteFile();
 
 		//! returns how much was written
-		virtual s32 write(const void* buffer, u32 sizeToWrite) _IRR_OVERRIDE_;
+		virtual size_t write(const void* buffer, size_t sizeToWrite) _IRR_OVERRIDE_;
 
 		//! changes position in file, returns true if successful
 		virtual bool seek(long finalPos, bool relativeMovement = false) _IRR_OVERRIDE_;
